Improved Cluster module interface.
It used to work with a class rather than instance. It now uses an instance, and now the interface also takes care of the management of available APIs, rather than outsource it in another file. All those changes make up for a much simpler code. It also now is able to be used with Facebook's submitit library, by making the interface pickable.